# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: Quermed CRM API
  version: 1.0.0
paths:
...
  /api/v1/me:
    get:
      tags:
      - me
      summary: Read the own profile and scope
      operationId: read_me_api_v1_me_get
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeRead'
    patch:
      tags:
      - me
      summary: Rename the own profile
      operationId: update_me_api_v1_me_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/today:
    get:
      tags:
      - me
      summary: 'The rep''s day: planned, overdue, week'
      operationId: read_today_api_v1_me_today_get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ActivityOutcome:
      type: string
      enum:
      - positive
      - neutral
      - negative
      - no_contact
      title: ActivityOutcome
    ActivityRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        account_id:
          type: string
          format: uuid
          title: Account Id
        account_name:
          type: string
          title: Account Name
        activity_type_id:
          type: string
          format: uuid
          title: Activity Type Id
        activity_type_name:
          type: string
          title: Activity Type Name
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        owner_name:
          type: string
          title: Owner Name
        status:
          $ref: '#/components/schemas/ActivityStatus'
        scheduled_at:
          type: string
          format: date-time
          title: Scheduled At
        done_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Done At
        duration_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Minutes
        outcome:
          anyOf:
          - $ref: '#/components/schemas/ActivityOutcome'
          - type: 'null'
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        cancel_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Cancel Reason
        opportunity_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Opportunity Id
        opportunity_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Opportunity Name
        contact_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Contact Ids
        contacts:
          items:
            $ref: '#/components/schemas/ContactNameRead'
          type: array
          title: Contacts
        next_activity_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Next Activity Id
        version:
          type: integer
          title: Version
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - account_id
      - account_name
      - activity_type_id
      - activity_type_name
      - owner_id
      - owner_name
      - status
      - scheduled_at
      - done_at
      - duration_minutes
      - outcome
      - subject
      - notes
      - cancel_reason
      - opportunity_id
      - opportunity_name
      - contact_ids
      - contacts
      - next_activity_id
      - version
      - created_at
      - updated_at
      title: ActivityRead
...
    ActivityStatus:
      type: string
      enum:
      - planned
      - done
      - cancelled
      title: ActivityStatus
...
    ContactNameRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: ContactNameRead
...
    DivisionRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        code:
          type: string
          title: Code
        name_es:
          type: string
          title: Name Es
        sort_order:
          type: integer
          title: Sort Order
      type: object
      required:
      - id
      - code
      - name_es
      - sort_order
      title: DivisionRead
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IdentityProvider:
      type: string
      enum:
      - password
      - entra_id
      title: IdentityProvider
...
    MeRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          $ref: '#/components/schemas/Role'
        is_active:
          type: boolean
          title: Is Active
        identity_provider:
          $ref: '#/components/schemas/IdentityProvider'
        territory_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Territory Ids
        division_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Division Ids
        version:
          type: integer
          title: Version
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        territories:
          items:
            $ref: '#/components/schemas/TerritoryRead'
          type: array
          title: Territories
        divisions:
          items:
            $ref: '#/components/schemas/DivisionRead'
          type: array
          title: Divisions
      type: object
      required:
      - id
      - email
      - full_name
      - role
      - is_active
      - identity_provider
      - territory_ids
      - division_ids
      - version
      - created_at
      - updated_at
      - territories
      - divisions
      title: MeRead
    MeUpdate:
      properties:
        full_name:
          type: string
          maxLength: 200
          minLength: 1
          title: Full Name
      type: object
      required:
      - full_name
      title: MeUpdate
...
    OpportunityStatus:
      type: string
      enum:
      - open
      - won
      - lost
      title: OpportunityStatus
    OpportunitySummaryRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        account_id:
          type: string
          format: uuid
          title: Account Id
        account_name:
          type: string
          title: Account Name
        name:
          type: string
          title: Name
        pipeline_id:
          type: string
          format: uuid
          title: Pipeline Id
        stage_id:
          type: string
          format: uuid
          title: Stage Id
        stage_name:
          type: string
          title: Stage Name
        division_id:
          type: string
          format: uuid
          title: Division Id
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        owner_name:
          type: string
          title: Owner Name
        status:
          $ref: '#/components/schemas/OpportunityStatus'
        amount:
          type: string
          title: Amount
        expected_close_date:
          type: string
          format: date
          title: Expected Close Date
        is_tender:
          type: boolean
          title: Is Tender
        tender_deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Tender Deadline
        is_at_risk:
          type: boolean
          title: Is At Risk
        stage_entered_at:
          type: string
          format: date-time
          title: Stage Entered At
        days_in_stage:
          type: integer
          title: Days In Stage
        version:
          type: integer
          title: Version
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - account_id
      - account_name
      - name
      - pipeline_id
      - stage_id
      - stage_name
      - division_id
      - owner_id
      - owner_name
      - status
      - amount
      - expected_close_date
      - is_tender
      - tender_deadline
      - is_at_risk
      - stage_entered_at
      - days_in_stage
      - version
      - updated_at
      title: OpportunitySummaryRead
...
    QuoteStatus:
      type: string
      enum:
      - draft
      - sent
      - accepted
      - rejected
      title: QuoteStatus
    QuoteSummaryRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        opportunity_id:
          type: string
          format: uuid
          title: Opportunity Id
        opportunity_name:
          type: string
          title: Opportunity Name
        account_id:
          type: string
          format: uuid
          title: Account Id
        account_name:
          type: string
          title: Account Name
        quote_number:
          type: string
          title: Quote Number
        display_number:
          type: string
          title: Display Number
        revision:
          type: integer
          title: Revision
        status:
          $ref: '#/components/schemas/QuoteStatus'
        total:
          type: string
          title: Total
        valid_until:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid Until
        is_expired:
          type: boolean
          title: Is Expired
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        owner_name:
          type: string
          title: Owner Name
        version:
          type: integer
          title: Version
        sent_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Sent At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - opportunity_id
      - opportunity_name
      - account_id
      - account_name
      - quote_number
      - display_number
      - revision
      - status
      - total
      - valid_until
      - is_expired
      - owner_id
      - owner_name
      - version
      - sent_at
      - created_at
      - updated_at
      title: QuoteSummaryRead
...
    Role:
      type: string
      enum:
      - sales_rep
      - sales_manager
      - back_office
      - admin
      title: Role
...
    TerritoryRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        provinces:
          items:
            type: string
          type: array
          title: Provinces
        is_active:
          type: boolean
          title: Is Active
        user_count:
          type: integer
          title: User Count
        version:
          type: integer
          title: Version
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - name
      - provinces
      - is_active
      - user_count
      - version
      - created_at
      - updated_at
      title: TerritoryRead
...
    TodayRead:
      properties:
        date:
          type: string
          format: date
          title: Date
        today:
          items:
            $ref: '#/components/schemas/ActivityRead'
          type: array
          title: Today
        overdue:
          items:
            $ref: '#/components/schemas/ActivityRead'
          type: array
          title: Overdue
        week:
          $ref: '#/components/schemas/WeekSummaryRead'
        tenders_due:
          items:
            $ref: '#/components/schemas/OpportunitySummaryRead'
          type: array
          title: Tenders Due
        at_risk:
          items:
            $ref: '#/components/schemas/OpportunitySummaryRead'
          type: array
          title: At Risk
        expiring_quotes:
          items:
            $ref: '#/components/schemas/QuoteSummaryRead'
          type: array
          title: Expiring Quotes
      type: object
      required:
      - date
      - today
      - overdue
      - week
      title: TodayRead
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WeekSummaryRead:
      properties:
        done_by_type:
          additionalProperties:
            type: integer
          propertyNames:
            format: uuid
          type: object
          title: Done By Type
        planned_remaining:
          type: integer
          title: Planned Remaining
      type: object
      required:
      - done_by_type
      - planned_remaining
      title: WeekSummaryRead
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer